+++ NEWS.md
@@ -15,8 +15,15 @@
 * Fixed index issue in 'mp_geocode'
 * Fixed 'mp_get_points' issue with multiple points
 
-# mapsapi 0.3.5
+# mapsapi 0.3.5 (2018-03-12)
 
 * Added 'bounds' parameter for 'mp_geocode'
 * Added argument validation in 'mp_geocode'
 * Corrected typo when creating geometry column in 'mp_get_routes' and 'mp_get_segments'
+
+# mapsapi 0.3.7
+
+* Minor updates in documentation
+* Added 'duration_in_traffic' option in 'mp_get_matrix'
+* 'mp_matrix' now returns 'NA' when there are zero results
+